% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nutrition_facts.R
\name{get_nutrition_facts}
\alias{get_nutrition_facts}
\title{Get nutrition facts}
\usage{
get_nutrition_facts(
  q,
  app_id = "1690fc80",
  app_key = "ea4faf1611e1ec77faf0b15cebc7356a",
  diet = "balanced",
  health = "",
  cuisineType = "",
  mealType = "",
  dishType = "",
  type = "public",
  beta = "false",
  public = "true"
)
}
\arguments{
\item{q}{A string}

\item{app_id}{A string}

\item{app_key}{A string}

\item{diet}{A string}

\item{health}{A string}

\item{cuisineType}{A string}

\item{mealType}{A string}

\item{dishType}{A string}

\item{type}{A string}

\item{beta}{A string}
}
\value{
A dataframe with recipe link and nutrition facts
}
\description{
Get nutrition facts
}
