% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe_info.R
\name{get_recipe_info}
\alias{get_recipe_info}
\title{Get recipe info}
\usage{
get_recipe_info(
  q,
  app_id = "1690fc80",
  app_key = "ea4faf1611e1ec77faf0b15cebc7356a"
)
}
\arguments{
\item{app_id}{a string}

\item{app_key}{a string}

\item{name}{a string}
}
\value{
A dataframe with the ingredients to a specific recipe
}
\description{
Get recipe info
}
