% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe_names.R
\name{get_recipe_names}
\alias{get_recipe_names}
\title{Get recipe names}
\usage{
get_recipe_names(
  q,
  app_id = "9f80d298",
  app_key = "e12319d6fd3db51d5a389c6336431579",
  diet = "balanced",
  health = "",
  cuisineType = "",
  mealType = "",
  dishType = "",
  type = "public",
  beta = "false",
  public = "true"
)
}
\arguments{
\item{q}{A string}

\item{app_id}{A string}

\item{app_key}{A string}

\item{diet}{A string}

\item{health}{A string}

\item{cuisineType}{A string}

\item{mealType}{A string}

\item{dishType}{A string}

\item{type}{A string}

\item{beta}{A string}
}
\value{
A dataframe with recipe names
}
\description{
Get recipe names
}
