% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUI.R
\name{shinyUI}
\alias{shinyUI}
\title{Gives you an appropriate side dish to accompany a main course}
\usage{
shinyUI(app_id = "6f567a3a", app_key = "aec97451eec00326ae7fedab93b7c250")
}
\arguments{
\item{app_id}{A string of your application id for edamam (required)}

\item{app_key}{A string of app key for edamam (required)}
}
\value{
shiny app
}
\description{
Gives you an appropriate side dish to accompany a main course
}
